Convert NoteEntryPage into a ContentView and move to new namespace
