+++ visitz/Visitz/VisitzConfig/VisitzScreens.cs
@@ -5,6 +5,7 @@ using Visitz.Views.Caseload;
 using Visitz.Views.Debugging;
 using Visitz.Views.Entity;
 using Visitz.Views.Navigation;
+using Visitz.Views.Notes;
 
 namespace Visitz.VisitzConfig
 {
@@ -38,7 +39,7 @@ namespace Visitz.VisitzConfig
             builder.Services.AddTransient<EntityNotesView>();
             builder.Services.AddTransient<EntityNotesViewModel>();
 
-            builder.Services.AddTransient<NoteEntryPage>();
+            builder.Services.AddTransient<NoteEntryView>();
             builder.Services.AddTransient<NoteEntryViewModel>();
 
             builder.Services.AddTransient<NotePublishPage>();
